Digital Ocean Update
This update:
Sets kubernetes version 1.5.7 as default on DO.
Allows for multi-master config on DO.
