+++ pkg/provider/digitalocean/provider.go
@@ -1,14 +1,12 @@
 package digitalocean
 
 import (
-	"bytes"
+	"io/ioutil"
+	"net/http"
 	"strconv"
-	"strings"
-	"text/template"
 	"time"
 
 	"github.com/digitalocean/godo"
-	"github.com/supergiant/supergiant/bindata"
 	"github.com/supergiant/supergiant/pkg/core"
 	"github.com/supergiant/supergiant/pkg/model"
 	"golang.org/x/oauth2"
@@ -25,218 +23,6 @@ func (p *Provider) ValidateAccount(m *model.CloudAccount) error {
 	client := p.Client(&model.Kube{CloudAccount: m})
 
 	_, _, err := client.Droplets.List(new(godo.ListOptions))
-	return err
-}
-
-// CreateKube creates a new DO kubernetes cluster.
-func (p *Provider) CreateKube(m *model.Kube, action *core.Action) error {
-	procedure := &core.Procedure{
-		Core:   p.Core,
-		Name:   "Create Kube",
-		Model:  m,
-		Action: action,
-	}
-
-	client := p.Client(m)
-
-	procedure.AddStep("creating global tags for Kube", func() error {
-		// These are created once, and then attached by name to created resource
-		globalTags := []string{
-			"Kubernetes-Cluster",
-			m.Name,
-			m.Name + "-master",
-			m.Name + "-minion",
-		}
-		for _, tag := range globalTags {
-			createInput := &godo.TagCreateRequest{
-				Name: tag,
-			}
-			if _, _, err := client.Tags.Create(createInput); err != nil {
-				// TODO
-				p.Core.Log.Warnf("Failed to create Digital Ocean tag '%s': %s", tag, err)
-			}
-		}
-		return nil
-	})
-
-	procedure.AddStep("creating master", func() error {
-		if m.MasterPublicIP != "" {
-			return nil
-		}
-
-		// Build template
-		masterUserdataTemplate, err := bindata.Asset("config/providers/digitalocean/master.yaml")
-		if err != nil {
-			return err
-		}
-		masterTemplate, err := template.New("master_template").Parse(string(masterUserdataTemplate))
-		if err != nil {
-			return err
-		}
-		var masterUserdata bytes.Buffer
-		if err = masterTemplate.Execute(&masterUserdata, m); err != nil {
-			return err
-		}
-
-		dropletRequest := &godo.DropletCreateRequest{
-			Name:              m.Name + "-master",
-			Region:            m.DigitalOceanConfig.Region,
-			Size:              m.MasterNodeSize,
-			PrivateNetworking: false,
-			UserData:          string(masterUserdata.Bytes()),
-			SSHKeys: []godo.DropletCreateSSHKey{
-				{
-					Fingerprint: m.DigitalOceanConfig.SSHKeyFingerprint,
-				},
-			},
-			Image: godo.DropletCreateImage{
-				Slug: "coreos-stable",
-			},
-		}
-		tags := []string{"Kubernetes-Cluster", m.Name, dropletRequest.Name}
-
-		masterDroplet, publicIP, err := p.createDroplet(client, action, dropletRequest, tags)
-		if err != nil {
-			return err
-		}
-
-		m.DigitalOceanConfig.MasterID = masterDroplet.ID
-		m.MasterPublicIP = publicIP
-		return nil
-	})
-
-	procedure.AddStep("building Kubernetes minion", func() error {
-		// Load Nodes to see if we've already created a minion
-		// TODO -- I think we can get rid of a lot of this do-unless behavior if we
-		// modify Procedure to save progess on Action (which is easy to implement).
-		if err := p.Core.DB.Find(&m.Nodes, "kube_name = ?", m.Name); err != nil {
-			return err
-		}
-		if len(m.Nodes) > 0 {
-			return nil
-		}
-
-		node := &model.Node{
-			KubeName: m.Name,
-			Kube:     m,
-			Size:     m.NodeSizes[0],
-		}
-		return p.Core.Nodes.Create(node)
-	})
-
-	// TODO repeated in provider_aws.go
-	procedure.AddStep("waiting for Kubernetes", func() error {
-		return action.CancellableWaitFor("Kubernetes API and first minion", 20*time.Minute, 3*time.Second, func() (bool, error) {
-			k8s := p.Core.K8S(m)
-			k8sNodes, err := k8s.ListNodes("")
-			if err != nil {
-				return false, nil
-			}
-			return len(k8sNodes) > 0, nil
-		})
-	})
-
-	return procedure.Run()
-}
-
-// DeleteKube deletes a DO kubernetes cluster.
-func (p *Provider) DeleteKube(m *model.Kube, action *core.Action) error {
-	// New Client
-	client := p.Client(m)
-	// Step procedure
-	procedure := &core.Procedure{
-		Core:   p.Core,
-		Name:   "Delete Kube",
-		Model:  m,
-		Action: action,
-	}
-
-	procedure.AddStep("deleting master", func() error {
-		if m.DigitalOceanConfig.MasterID == 0 {
-			return nil
-		}
-		if _, err := client.Droplets.Delete(m.DigitalOceanConfig.MasterID); err != nil && !strings.Contains(err.Error(), "404") {
-			return err
-		}
-		m.DigitalOceanConfig.MasterID = 0
-		return nil
-	})
-
-	return procedure.Run()
-}
-
-// CreateNode creates a new minion on DO kubernetes cluster.
-func (p *Provider) CreateNode(m *model.Node, action *core.Action) error {
-	// Build template
-	minionUserdataTemplate, err := bindata.Asset("config/providers/digitalocean/minion.yaml")
-	if err != nil {
-		return err
-	}
-	minionTemplate, err := template.New("master_template").Parse(string(minionUserdataTemplate))
-	if err != nil {
-		return err
-	}
-
-	data := struct {
-		*model.Node
-		Token string
-	}{
-		m,
-		m.Kube.CloudAccount.Credentials["token"],
-	}
-
-	var minionUserdata bytes.Buffer
-	if err = minionTemplate.Execute(&minionUserdata, data); err != nil {
-		return err
-	}
-
-	dropletRequest := &godo.DropletCreateRequest{
-		Name:              m.Kube.Name + "-minion",
-		Region:            m.Kube.DigitalOceanConfig.Region,
-		Size:              m.Size,
-		PrivateNetworking: true,
-		UserData:          string(minionUserdata.Bytes()),
-		SSHKeys: []godo.DropletCreateSSHKey{
-			{
-				Fingerprint: m.Kube.DigitalOceanConfig.SSHKeyFingerprint,
-			},
-		},
-		Image: godo.DropletCreateImage{
-			Slug: "coreos-stable",
-		},
-	}
-	tags := []string{"Kubernetes-Cluster", m.Kube.Name, dropletRequest.Name}
-
-	minionDroplet, publicIP, err := p.createDroplet(p.Client(m.Kube), action, dropletRequest, tags)
-	if err != nil {
-		return err
-	}
-
-	// Parse creation timestamp
-	createdAt, err := time.Parse("2006-01-02T15:04:05Z", minionDroplet.Created)
-	if err != nil {
-		// TODO need to return on error here
-		p.Core.Log.Warnf("Could not parse Droplet creation timestamp string '%s': %s", minionDroplet.Created, err)
-	}
-
-	// Save info before waiting on IP
-	m.ProviderID = strconv.Itoa(minionDroplet.ID)
-	m.ProviderCreationTimestamp = createdAt
-	m.ExternalIP = publicIP
-	m.Name = publicIP
-
-	return p.Core.DB.Save(m)
-}
-
-// DeleteNode deletes a minsion on a DO kubernetes cluster.
-func (p *Provider) DeleteNode(m *model.Node, action *core.Action) error {
-	client := p.Client(m.Kube)
-
-	intID, err := strconv.Atoi(m.ProviderID)
-	if err != nil {
-		return err
-	}
-	_, err = client.Droplets.Delete(intID)
 	return err
 }
 
@@ -316,4 +102,18 @@ func (p *Provider) createDroplet(client *godo.Client, action *core.Action, req *
 	}
 
 	return droplet, publicIP, nil
+}
+
+func etcdToken(num string) (string, error) {
+	resp, err := http.Get("https://discovery.etcd.io/new?size=" + num + "")
+	if err != nil {
+		return "", err
+	}
+	defer resp.Body.Close()
+
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return "", err
+	}
+	return string(body), nil
 }
